u-boot: Enable autoboot stop string functionality
Enable the autoboot stop string functionality of the u-boot by default.
This in practice allows defining a password that can be used to lock the
command line interface of the u-boot. If no password is given, the
command line interface becomes inaccessible.

To enter the password, user first has to press enter to stop the
autoboot, then enter the password, and then press enter again.

This feature makes the bootloader more secure, as it should not be
possible to enter the command line without knowing the password.