u-boot: Prevent reading console password from environment
The default functionality of the u-boot is to read bootstopkeycrypt
variable from the environment for the password that is used to unlock
the command line interface. However, trusting this environment variable
is dangerous, as the adversary may be able to modify and there is no
integrity checking for environment. It becomes fairly easy for an
attacker to add (or modify) the variable with a known value and get an
access.

Remove the usage of the bootstopkeycrypt environment value, and always
use the built-in value. Modifying this is possible as well, but should
be slightly more difficult.