u-boot: Add unlocking feature
Add an unlocking feature to the u-boot. This feature can be used to set
the device into unlocked state if some condition is met, and the status
of the locking can be checked.

As an initial implementation for Sulka, set the u-boot into an unlocked
state if the crypted stop string is input correctly and the autoboot is
aborted. This allows privileged command line interface, but should limit
the functionality of the autoboot that cannot be aborted without knowing
the password.