u-boot: Prevent CLI on failed autoboot
By default, u-boot drops to the CLI if autoboot fails. However, this can
be a security issue if someone manages to modify the autoboot process so
that the boot fails. Do not drop into the CLI if the device is still in
locked state. In the current Sulka approach where the device can be
unlocked only with autoboot stop string this should always be false, but
in future this may change.