u-boot: Implement command allowlist feature
Implement a feature that can be used to allowlist commands in the
u-boot. The feature allows limiting commands if the device is in locked
state, and executing all the available commands if the device is in
unlocked state. This can be used to secure the autoboot flow from
unexpected commands.

Note that the unnecessary commands should still be removed from the
u-boot configuration, as they are accessible in the unlocked CLI and can
pose a security risk.

The feature is disabled by default, as it is impossible from the distro
point of view to know what commands should be whitelisted, and the
command list varies between different devices, use cases, etc. To enable
the feature, add the following to the u-boot configuration:

CONFIG_COMMAND_ALLOWLIST=y
CONFIG_COMMAND_ALLOWLIST_CMDS="space separated list of allowed commands"